pyproject.toml
[project]
name = "bigocrpdf"
version = "3.0.0"
description = "Add OCR to your PDF documents to make them searchable - Powered by RapidOCR"
readme = "README.md"
requires-python = ">=3.10"
license = "GPL-3.0-or-later"
authors = [
    {name = "BigLinux Team", email = "contact@biglinux.com.br"}
]
keywords = ["ocr", "pdf", "gtk", "libadwaita", "biglinux"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: X11 Applications :: GTK",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Utilities",
]
dependencies = [
    "PyGObject>=3.42.0",
    "pycairo>=1.20.0",
    # RapidOCR engine
    "rapidocr>=3.5.0",
    # PDF processing
    "pikepdf>=8.0.0",
    "reportlab>=4.0.0",
    # Image processing
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    # Dewarp dependencies
    "scipy>=1.11.0",
    # ODF export
    "odfpy>=1.4.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "ruff"]



[project.scripts]
bigocrpdf-cli = "bigocrpdf.cli:main"

[project.gui-scripts]
bigocrpdf = "bigocrpdf:main"

[project.urls]
Homepage = "https://www.biglinux.com.br"
Repository = "https://github.com/biglinux/bigocrpdf"
Issues = "https://github.com/biglinux/bigocrpdf/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
bigocrpdf = ["resources/*.css"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "C4", "UP"]
ignore = ["E501", "E402"]  # E402: Module level import not at top (needed for gi.require_version)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
pythonpath = ["src"]
